Move Projects files out of main app
