+++ VoiceOver Designer/Documents/VODesignDocument+Window.swift
@@ -1,0 +1,8 @@
+import Document
+
+extension VODesignDocument {
+    public override func makeWindowControllers() {
+        WindowManager.shared
+            .createNewDocumentWindow(document: self)
+    }
+}
